m-automaticontratos-api/requirements.txt
connexion == 2.14.2
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
gunicorn==21.2.0
six~=1.16.0
Flask==2.2.5
requests~=2.32.3
uuid==1.30
APScheduler==3.9.1.post1
Flask-Cors==4.0.0
loguru~=0.7.2
pandas~=2.0.3
openpyxl~=3.1.5
cx-Oracle~=8.3.0
Booktype~=1.5
mysql-connector-python~=9.0.0